requirements.txt
transformers==4.53.2
torch==2.6.0
protobuf==3.20.3
praw==7.8.1
python-dotenv==1.1.1
supabase==2.17.0
